Bug 645967: removed build - r=mixedpuppy,gozer
