+++ build.py
@@ -33,71 +33,141 @@
 # the terms of any one of the MPL, the GPL or the LGPL.
 #
 # ***** END LICENSE BLOCK *****
-"""
-Bootstrap file -- will try to make sure _build.py is up-to-date, then run it.
-"""
 import os
 import sys
-import urllib2
+import subprocess
 
 
-def _rename(path):
-    if not os.path.exists(path):
-        return
-
-    root = 0
-    newname = path + '.bak.%d' % root
-    while os.path.exists(newname):
-        root += 1
-        newname = path + '.bak.%d' % root
-    os.rename(path, newname)
+CURDIR = os.path.dirname(__file__)
+REPO_ROOT = 'https://github.com/mozilla/%s.git'
+PYTHON = sys.executable
 
 
-def _update():
-    # getting the file age
-    if os.path.exists('._build.etag'):
-        with open('._build.etag') as f:
-            current_etag = f.read().strip()
-        headers = {'If-None-Match': current_etag}
+def _get_tags():
+    sub = subprocess.Popen('git tag', shell=True, stdout=subprocess.PIPE)
+    tags = [line.strip()
+            for line in sub.stdout.read().strip().split('\n')
+            if line.strip() != '']
+    tags.reverse()
+    return tags
+
+
+def verify_tag(tag):
+    if tag == 'tip' or tag.isdigit():
+        return True
+    return tag in _get_tags()
+
+
+def get_latest_tag():
+    tags = _get_tags()
+    if len(tags) == 0:
+        raise ValueError('Could not find a tag')
+    return tags[0]
+
+
+def _run(command):
+    print(command)
+    os.system(command)
+
+
+def _envname(name):
+    return name.upper().replace('-', '_')
+
+
+def _update_cmd(project, latest_tags=False):
+    if latest_tags:
+        return 'git checkout -r "%s"' % get_latest_tag()
     else:
-        headers = {}
-        current_etag = None
 
-    request = urllib2.Request('http://moz.ziade.org/_build.py',
-                              headers=headers)
+        # looking for an environ with a specific tag or rev
+        rev = os.environ.get(_envname(project))
+        if rev is not None:
 
-    # checking the last version on our server
+            if not verify_tag(rev):
+                print('Unknown tag or revision: %s' % rev)
+                sys.exit(1)
+
+            return 'git checkout -r "%s"' % rev
+        return 'git checkout'
+
+
+def build_app(name, latest_tags, deps):
+    # building deps first
+    build_deps(deps, latest_tags)
+
+    # build the app now
+    if not _has_spec():
+        latest_tags = False
+
+    _run(_update_cmd(name, latest_tags))
+    _run('%s setup.py develop' % PYTHON)
+
+
+def build_deps(deps, latest_tags):
+    """Will make sure dependencies are up-to-date"""
+    location = os.getcwd()
+    # do we want the latest tags ?
     try:
-        url = urllib2.urlopen(request, timeout=5)
-        etag = url.headers.get('ETag')
-    except urllib2.HTTPError, e:
-        if e.getcode() != 412:
-            raise
-        # we're up-to-date (precondition failed)
-        etag = current_etag
-    except urllib2.URLError:
-        # timeout error
-        etag = None
+        deps_dir = os.path.abspath(os.path.join(CURDIR, 'deps'))
+        if not os.path.exists(deps_dir):
+            os.mkdir(deps_dir)
 
-    if etag is not None and current_etag != etag:
-        # we need to update our version
-        _rename('_build.py')
-        content = url.read()
-        with open('_build.py', 'w') as f:
-            f.write(content)
+        for dep in deps:
+            repo = REPO_ROOT % dep
+            target = os.path.join(deps_dir, dep)
+            if os.path.exists(target):
+                os.chdir(target)
+                _run('git pull')
+            else:
+                _run('git clone %s %s' % (repo, target))
+                os.chdir(target)
 
-        with open('._build.etag', 'w') as f:
-            f.write(etag)
+            update_cmd = _update_cmd(dep, latest_tags)
+            _run(update_cmd)
+            _run('%s setup.py develop' % PYTHON)
+    finally:
+        os.chdir(location)
 
 
-def main():
-    #_update()
-    # we're good, let's import the file and run it
-    mod = __import__('_build')
-    project_name = sys.argv[1]
-    deps = [dep.strip() for dep in sys.argv[2].split(',')]
-    mod.main(project_name, deps)
+def _has_spec():
+    specs = [file_ for file_ in os.listdir('.')
+             if file_.endswith('.spec')]
+    return len(specs)
+
+
+def main(project_name, deps):
+    # check the provided values in the environ
+    latest_tags = 'LATEST_TAGS' in os.environ
+
+    if not latest_tags:
+        # if we have some tags in the environ, check that they are all defined
+        projects = list(deps)
+
+        # is the root a project itself or just a placeholder ?
+        if _has_spec():
+            projects.append(project_name)
+
+        tags = {}
+        missing = 0
+        for project in projects:
+            tag = _envname(project)
+            if tag in os.environ:
+                tags[tag] = os.environ[tag]
+            else:
+                tags[tag] = 'Not provided'
+                missing += 1
+
+        # we want all tag or no tag
+        if missing > 0 and missing < len(projects):
+            print("You did not specify all tags: ")
+            for project, tag in tags.items():
+                print('    %s: %s' % (project, tag))
+            sys.exit(1)
+
+    build_app(project_name, latest_tags, deps)
 
 
 if __name__ == '__main__':
-    main()
+    project_name = sys.argv[1]
+    deps = [dep.strip() for dep in sys.argv[2].split(',')]
+    main(project_name, deps)
